refactor hooks dir
