+++ src/hooks/tts-hooks/use-elevenlabs.ts
@@ -1,9 +1,9 @@
 // hooks/useTextToSpeech.ts
 import { useCallback, useEffect, useState } from 'react';
 import axios from 'axios';
-import { getToken } from '../utils/config';
-import { Option } from '../components/DropDown/';
-import { useLocalStorageKeyValue } from './use-local-storage';
+import { getToken } from '../../utils/config';
+import { Option } from '../../components/DropDown';
+import { useLocalStorageKeyValue } from '../use-local-storage';
 
 export const useElevenlabs = () => {
   const [voices, setVoices] = useState<Option[]>([]);
